requirements.txt
anyio==3.6.2
APScheduler==3.9.1
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
et-xmlfile==1.1.0
gTTS==2.2.4
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
imutils==0.5.4
libretranslatepy==2.1.1
lxml==4.9.1
numpy==1.23.4
opencv-python==4.6.0.66
openpyxl==3.0.10
pandas==1.5.1
playsound==1.3.0
psycopg2-binary==2.9.5
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==20.0a4
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-futures==1.0.0
result==0.8.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stem==1.8.1
tornado==6.2
torrequest==0.1.0
translate==3.6.1
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
wikipedia==1.4.0
